"Transaction update functionality when amount changes"

+++ finance-tracker-backend/src/main/java/com/finance/tracker/accounts/domain/SnapshotCreateRequest.java
@@ -11,8 +11,8 @@ public class SnapshotCreateRequest {
 
     private String accountId;
     private String transactionId;
-    private BigDecimal previousBalance;
-    private BigDecimal newBalance;
-    private Double transactionAmount;
+    private double previousBalance;
+    private double newBalance;
+    private double transactionAmount;
 
 }
